a few more things to util
