+++ corehq/apps/userreports/sql/adapter.py
@@ -1,8 +1,8 @@
-import hashlib
 import sqlalchemy
 from sqlalchemy.exc import IntegrityError, ProgrammingError
 from corehq.apps.userreports.exceptions import TableRebuildError
 from corehq.apps.userreports.sql.columns import column_to_sql
+from corehq.apps.userreports.sql.util import get_table_name
 from dimagi.utils.decorators.memoized import memoized
 
 
@@ -73,10 +73,3 @@ def rebuild_table(engine, table):
         table.drop(connection, checkfirst=True)
         table.create(connection)
     engine.dispose()
-
-
-def get_table_name(domain, table_id):
-    def _hash(domain, table_id):
-        return hashlib.sha1('{}_{}'.format(hashlib.sha1(domain).hexdigest(), table_id)).hexdigest()[:8]
-
-    return 'config_report_{0}_{1}_{2}'.format(domain, table_id, _hash(domain, table_id))
