+++ corehq/apps/userreports/sql/columns.py
@@ -1,4 +1,3 @@
-import hashlib
 import sqlalchemy
 from sqlagg import SumWhen
 from corehq.db import Session
@@ -58,27 +57,6 @@ def get_expanded_column_config(data_source_configuration, column_config, lang):
             )
         )
     return SqlColumnConfig(_expand_column(column_config, vals, lang), warnings=column_warnings)
-
-
-def get_column_name(path):
-    """
-    :param path: xpath from form or case
-    :return: column name for postgres
-
-    Postgres only allows columns up to 63 characters
-    Anyone viewing the table directly will want to know the last parts of the path, not the first parts e.g.
-    this: 'my_long_choice_list_option_1_ABCDEFGH', 'my_long_choice_list_option_2_ABCD1234'
-    not: 'question_group_1_my_long_choice_ABCDEFGH', 'question_group_1_my_long_choice_ABCD1234'
-    """
-    parts = path.split("/")
-
-    def _hash(parts):
-        front = "/".join(parts[:-1])
-        end = parts[-1]
-        return hashlib.sha1('{}_{}'.format(hashlib.sha1(front).hexdigest(), end)).hexdigest()[:8]
-
-    new_parts = path[-54:].split("/")
-    return "_".join(new_parts + [_hash(parts)])
 
 
 def _get_distinct_values(data_source_configuration, column_config, expansion_limit=10):
